backend/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
psycopg2-binary
youtube-transcript-api
openai
pydantic-settings
python-dotenv
langchain-text-splitters
